ee/requirements.txt
torch==1.11.0
bertviz==1.4.0
blinker==1.4
brotlipy==0.7.0
conllu==4.4.1
dominate==2.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
en-dual-none-contextual @ https://github.com/UCREL/pymusas-models/releases/download/en_dual_none_contextual-0.3.1/en_dual_none_contextual-0.3.1-py3-none-any.whl
fonttools==4.25.0
google-auth-oauthlib==0.4.1
munkres==1.1.4
nmslib==2.1.1
pandas==1.4.1
patsy==0.5.2
Pillow==9.0.1
protobuf==3.19.1
pyasn1-modules==0.2.8
pybind11==2.6.1
pymusas==0.3.0
pysbd==0.3.4
pytz==2021.3
PyYAML==6.0
requests-oauthlib==1.3.0
sadice==0.1.3
scispacy==0.5.0
sentencepiece==0.1.97
spacy==3.2.4
spacy-alignments==0.8.5
spacy-transformers==1.1.5
